GitHub/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ma-migration"
version = "0.1.0"
description = "Generalized, domain-first data preparation library refactored from Customers_Breakdown notebook."
requires-python = ">=3.10"
dependencies = ["pandas>=2.0", "numpy>=1.24", "python-dateutil>=2.8", "openpyxl>=3.1", "pyyaml>=6.0"]

[tool.setuptools.packages.find]
where = ["src"]
